Update API to use search-client crate
