+++ medicines/api/src/product.rs
@@ -1,5 +1,6 @@
-use crate::azure_search::{AzureResult, AzureSearchClient};
 use juniper::GraphQLObject;
+use search_client::models::AzureResult;
+use search_client::Search;
 
 #[derive(GraphQLObject, Eq, Ord, PartialEq, PartialOrd)]
 #[graphql(description = "A medical product containing active ingredients")]
@@ -34,24 +35,20 @@ pub fn handle_doc(document: &AzureResult, products: &mut Vec<Product>) {
 }
 
 pub async fn get_products_by_substance_name(
-    substance_name: String,
-    client: &AzureSearchClient,
+    substance_name: &str,
+    client: &impl Search,
 ) -> Vec<Product> {
     // Get a list of documents from Azure which are about products containing the
     // substance name.
     let azure_result = client
-        .filter_by_collection(
-            "substance_name".to_string(),
-            substance_name,
-            "eq".to_string(),
-        )
+        .filter_by_field("substance_name", substance_name)
         .await
         .unwrap();
 
     // Extract a list of products while keeping track of the number of documents that
     // product has.
     let mut products = Vec::<Product>::new();
-    for document in azure_result.value {
+    for document in azure_result.search_results {
         handle_doc(&document, &mut products);
     }
 
@@ -68,7 +65,7 @@ mod test {
         AzureResult {
             product_name: product_name,
             doc_type: "dummy".to_string(),
-            created: "yes".to_string(),
+            created: Some("yes".to_string()),
             facets: Vec::new(),
             file_name: "README.markdown".to_string(),
             highlights: None,
@@ -76,7 +73,7 @@ mod test {
             metadata_storage_name: "dummy".to_string(),
             metadata_storage_path: "/".to_string(),
             metadata_storage_size: 0,
-            release_state: "solid".to_string(),
+            release_state: Some("solid".to_string()),
             rev_label: None,
             score: -0.0,
             substance_name: Vec::new(),
